requirements.txt
tensorflow-gpu
scipy
matplotlib
ipython
opencv-python
scikit-image
pyyaml
h5py
